Use a JS script to redirect the page to the login screen if no user is currently logged in.
If no user is currently logged in on the view issue or create issue
screens then a script is added to the page header that redirects the
browser to the login page.

The screens where redirection may occur is controlled by context tags on
the web-resource in the plugin's atlassian-plugin.xml.

NB: Despite its name, the builtin
com.atlassian.jira.plugin.webfragment.conditions.UserLoggedInCondition
class is not suitable to be used by this plugin.